Deleting the old search component
