+++ app/src/app/components/search/SearchComponent.css
