+++ app/src/app/components/search/SearchContainer.ts
@@ -1,9 +1,9 @@
 import { connect } from 'react-redux';
 import SearchComponent from './SearchComponent';
-import { IStoreState } from '../../../types';
+import { IStoreState } from '../../types';
 import { Dispatch } from 'redux';
-import { actions, MicroPadAction } from '../../../actions';
-import { RestoreJsonNotepadAndLoadNoteAction } from '../../../types/ActionTypes';
+import { actions, MicroPadAction } from '../../actions';
+import { RestoreJsonNotepadAndLoadNoteAction } from '../../types/ActionTypes';
 
 export const searchConnector = connect(
 	({ search, notepads }: IStoreState) => {
